refactor: reorganize frontend types into base/ and events/ folders
- Move primitive types (Action, Observation, base interfaces, common types) to base/ folder
- Move main event types (ActionEvent, MessageEvent, etc.) to events/ folder
- Update all import statements to reflect new folder structure
- Add index.ts files for cleaner imports
- Improve type organization and maintainability
