+++ frontend/src/types/v1/core/base/base.ts
