+++ frontend/src/types/v1/core/base/common.ts
