+++ frontend/src/types/v1/core/base/event.ts
