+++ frontend/src/types/v1/core/base/index.ts
@@ -1,0 +1,6 @@
+// Export all base types
+export * from "./action";
+export * from "./base";
+export * from "./common";
+export * from "./event";
+export * from "./observation";
