+++ frontend/src/types/v1/core/events/pause-event.ts
@@ -1,4 +1,4 @@
-import { BaseEvent } from "./event";
+import { BaseEvent } from "../base/event";
 
 // Pause event - indicates that agent execution was paused by user request
 export interface PauseEvent extends BaseEvent {
