+++ frontend/src/types/v1/core/index.ts
@@ -1,13 +1,7 @@
 // Export all core types
-export * from "./action";
-export * from "./action-event";
+
+// Base types (primitive types like Action, Observation, common interfaces)
 export * from "./base";
-export * from "./common";
-export * from "./condensation-event";
-export * from "./conversation-state-event";
-export * from "./event";
-export * from "./message-event";
-export * from "./observation";
-export * from "./observation-event";
-export * from "./pause-event";
-export * from "./system-event";
+
+// Event types (main events that extend BaseEvent)
+export * from "./events";
